Feature controller done
Register || update || delete || update
